+++ plugin_ide.ui/src-lang/melnorme/util/swt/components/fields/EnablementButtonTextField.java
@@ -66,7 +66,7 @@ public abstract class EnablementButtonTextField extends ButtonTextField {
 			try {
 				setFieldValue(getDefaultFieldValue());
 			} catch(CommonException e) {
-				// Do nothing;
+				setFieldValue("");
 			}
 		}
 	}