+++ documentation/ChangeLog.md
@@ -2,6 +2,8 @@
 [Latest features on top]
 
 ### (NextVersion)
+ * Fixed: project Build Targets settings pages shows wrong default for Program Path field
+ 
  * Syntax highlighting now works on the source compare editor/viewer.
  * Added per-project compiler/SDK installation preferences.
  * Newly created launch configurations now have the debug option "Stop on startup at:" set to false by default. This way debugging won't stop on the C `main`, which is essentially useless outside of C/C++.